requirements.txt
Django==1.10.5
djangorestframework==3.5.3
pytest-django==3.1.2